+++ contrib/dopus/Program/main17.c
@@ -591,10 +591,10 @@ char *command;
             if ((entry=findfile(dopus_curwin[win],rexx_args[0],NULL))) {
                 getprot(entry->protection,buf2);
                 lsprintf(buf,"%s%lc%ld%lc%ld%lc%ld%lc%ld%lc%ld%lc%ld%lc%s%lc%s%lc",
-                    entry->name,pad[0],(int)entry->size,pad[0],entry->subtype,pad[0],
-                    entry->type,pad[0],entry->selected,pad[0],
-                    entry->date.ds_Days,pad[0],
-                    (entry->date.ds_Minute*60)+(entry->date.ds_Tick/50),pad[0],
+                    entry->name,pad[0],(long int)entry->size,pad[0],(long int)entry->subtype,pad[0],
+                    (long int)entry->type,pad[0],(long int)entry->selected,pad[0],
+                    (long int)entry->date.ds_Days,pad[0],
+                    (long int)(entry->date.ds_Minute*60)+(entry->date.ds_Tick/50),pad[0],
                     entry->comment,pad[0],buf2,pad[0]);
                 rexx_set_return(msg,0,buf);
                 return(1);
@@ -920,7 +920,7 @@ doentryselection:
                         if (b<0 || b>1) b=data_active_window;
                         for (a=0;a<=DISPLAY_LAST;a++) {
                             if (config->displaypos[b][a]==-1) break;
-                            lsprintf(buf2,"%ld ",config->displaypos[b][a]);
+                            lsprintf(buf2,"%ld ",(long int)config->displaypos[b][a]);
                             strcat(buf,buf2);
                         }
                         retbuf=buf;
@@ -1012,7 +1012,7 @@ doentryselection:
 
                     case MOD_FONT:
                         if (b<0 || b>=FONT_LAST) break;
-                        lsprintf(buf,"%s/%ld",config->fontbufs[b],config->fontsizes[b]);
+                        lsprintf(buf,"%s/%ld",config->fontbufs[b],(long int)config->fontsizes[b]);
                         retbuf=buf;
                         break;
 
@@ -1034,11 +1034,11 @@ doentryselection:
                         buf[0]=0;
                         if (b<0 || b>1) b=data_active_window;
                         lsprintf(buf,"%ld %ld %ld %ld %ld",
-                            config->displaylength[b][DISPLAY_NAME],
-                            config->displaylength[b][DISPLAY_COMMENT],
-                            config->displaylength[b][DISPLAY_FILETYPE],
-                            config->displaylength[b][DISPLAY_OWNER],
-                            config->displaylength[b][DISPLAY_GROUP]);
+                            (long int)config->displaylength[b][DISPLAY_NAME],
+                            (long int)config->displaylength[b][DISPLAY_COMMENT],
+                            (long int)config->displaylength[b][DISPLAY_FILETYPE],
+                            (long int)config->displaylength[b][DISPLAY_OWNER],
+                            (long int)config->displaylength[b][DISPLAY_GROUP]);
                         retbuf=buf;
                         break;
 
@@ -1062,15 +1062,15 @@ doentryselection:
 
                     case MOD_WINDOWXY:
                         if (!status_iconified && status_publicscreen && Window)
-                            lsprintf(buf,"%ld %ld",Window->LeftEdge,Window->TopEdge);
-                        else lsprintf(buf,"%ld %ld",config->wbwinx,config->wbwiny);
+                            lsprintf(buf,"%ld %ld",(long int)Window->LeftEdge,(long int)Window->TopEdge);
+                        else lsprintf(buf,"%ld %ld",(long int)config->wbwinx,(long int)config->wbwiny);
                         retbuf=buf;
                         break;
 
                     case MOD_WINDOWSIZE:
                         if (!status_iconified && status_publicscreen && Window)
-                            lsprintf(buf,"%ld %ld",Window->Width,Window->Height);
-                        else lsprintf(buf,"%ld %ld",config->scr_winw,config->scr_winh);
+                            lsprintf(buf,"%ld %ld",(long int)Window->Width,(long int)Window->Height);
+                        else lsprintf(buf,"%ld %ld",(long int)config->scr_winw,(long int)config->scr_winh);
                         retbuf=buf;
                         break;
 
@@ -1081,11 +1081,11 @@ doentryselection:
                     case MOD_WINDOWXYWH:
                         if (!status_iconified && status_publicscreen && Window) {
                             lsprintf(buf,"%ld %ld %ld %ld",
-                                Window->LeftEdge,Window->TopEdge,Window->Width,Window->Height);
+                                (long int)Window->LeftEdge,(long int)Window->TopEdge,(long int)Window->Width,(long int)Window->Height);
                         }
                         else {
                             lsprintf(buf,"%ld %ld %ld %ld",
-                                config->wbwinx,config->wbwiny,config->scr_winw,config->scr_winh);
+                                (long int)config->wbwinx,(long int)config->wbwiny,(long int)config->scr_winw,(long int)config->scr_winh);
                         }
                         retbuf=buf;
                         break;
